// frontend/src/app/api/model/user-project-response.ts
/**
 * Experiment Tracking Platform API
 *
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

export interface UserProjectResponse {
  projectId?: string;
  name?: string;
  description?: string;
  createdAt?: string;
  role?: UserProjectResponse.RoleEnum;
  joinedAt?: string;
}
export namespace UserProjectResponse {
  export const RoleEnum = {
    OWNER: 'OWNER',
    EDITOR: 'EDITOR',
    VIEWER: 'VIEWER',
  } as const;
  export type RoleEnum = (typeof RoleEnum)[keyof typeof RoleEnum];
}